WIP: Attempt tail call optimisation. Not working!
Passing around a boolean indicating when we are in a tail context in the Scheme function.
However we never end up using it for anything!
Key thing seems to be to make sure Haskell itself never have a stack overflow
One way to help is to ensure strictness rather than laziness (not sure what this involves)
In this commit, eval and apply use tail-calls but not tail-*recursive* calls.
i.e. Return value of function is a call to *another* function
Worth trying to refactor for tail recursion, then try strict/lazy stuff.